Implemented TypeScript to docs.
